backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.1.0
djangorestframework-simplejwt==5.3.0
django-filter==23.5
reportlab==4.0.7
stripe==7.8.0
django-storages==1.14.2
boto3==1.34.10
celery==5.3.4
redis==5.0.1
social-auth-app-django==5.4.0
